+++ RealChute/EVA/EVAChuteLibrary.cs
@@ -11,7 +11,7 @@ using UnityEngine;
  * inactive (no connection) for a period of 90 days on the official KSP forums. In that case, the license reverts
  * back to CC-BY-NC-SA 4.0 INTL.*/
 
-namespace RealChute.Libraries
+namespace RealChute.EVA
 {
     public class EVAChuteLibrary
     {
