+++ RealChute/EVA/EVAChuteTest.cs
@@ -10,7 +10,7 @@ using RealChute.Extensions;
  * inactive (no connection) for a period of 90 days on the official KSP forums. In that case, the license reverts
  * back to CC-BY-NC-SA 4.0 INTL.*/
 
-namespace RealChute
+namespace RealChute.EVA
 {
     public class EVAChuteTest : PartModule
     {
