+++ RealChute/Extensions/CelestialBodyExtensions.cs
@@ -1,4 +1,5 @@
 ﻿using System;
+using RealChute.Utils;
 using Debug = UnityEngine.Debug;
 
 /* RealChute was made by Christophe Savard (stupid_chris). You are free to copy, fork, and modify RealChute as you see
