+++ RealChute/Extensions/ConfigNodeExtensions.cs
@@ -1,4 +1,5 @@
-﻿using UnityEngine;
+﻿using RealChute.Utils;
+using UnityEngine;
 
 /* RealChute was made by Christophe Savard (stupid_chris). You are free to copy, fork, and modify RealChute as you see
  * fit. However, redistribution is only permitted for unmodified versions of RealChute, and under attribution clause.
