+++ RealChute/KerbalDumper.cs
@@ -1,6 +1,7 @@
 ﻿using System.Text;
-using UnityEngine;
 using RealChute.Extensions;
+using RealChute.Utils;
+using UnityEngine;
 
 /* RealChute was made by Christophe Savard (stupid_chris). You are free to copy, fork, and modify RealChute as you see
  * fit. However, redistribution is only permitted for unmodified versions of RealChute, and under attribution clause.
