+++ RealChute/Libraries/Materials/MaterialDefinition.cs
@@ -9,7 +9,7 @@
  * inactive (no connection) for a period of 90 days on the official KSP forums. In that case, the license reverts
  * back to CC-BY-NC-SA 4.0 INTL.*/
 
-namespace RealChute.Libraries
+namespace RealChute.Libraries.Materials
 {
     public class MaterialDefinition
     {
