+++ RealChute/Libraries/Materials/MaterialsLibrary.cs
@@ -13,7 +13,7 @@ using RealChute.Extensions;
  * inactive (no connection) for a period of 90 days on the official KSP forums. In that case, the license reverts
  * back to CC-BY-NC-SA 4.0 INTL.*/
 
-namespace RealChute.Libraries
+namespace RealChute.Libraries.Materials
 {
     public class MaterialsLibrary
     {
