+++ RealChute/Libraries/Presets/Preset.cs
@@ -1,6 +1,9 @@
 ﻿using System.Collections.Generic;
 using System.Linq;
 using RealChute.Extensions;
+using RealChute.GUI;
+using RealChute.Utils;
+using RealChute.Libraries.Materials;
 
 /* RealChute was made by Christophe Savard (stupid_chris). You are free to copy, fork, and modify RealChute as you see
  * fit. However, redistribution is only permitted for unmodified versions of RealChute, and under attribution clause.
@@ -11,7 +14,7 @@ using RealChute.Extensions;
  * inactive (no connection) for a period of 90 days on the official KSP forums. In that case, the license reverts
  * back to CC-BY-NC-SA 4.0 INTL.*/
 
-namespace RealChute.Libraries
+namespace RealChute.Libraries.Presets
 {
     public class Preset
     {
