+++ RealChute/Libraries/Presets/PresetsLibrary.cs
@@ -13,7 +13,7 @@ using UnityEngine;
  * inactive (no connection) for a period of 90 days on the official KSP forums. In that case, the license reverts
  * back to CC-BY-NC-SA 4.0 INTL.*/
 
-namespace RealChute.Libraries
+namespace RealChute.Libraries.Presets
 {
     public class PresetsLibrary
     {
