+++ RealChute/Libraries/Textures/ModelConfig.cs
@@ -11,7 +11,7 @@ using RealChute.Extensions;
  * inactive (no connection) for a period of 90 days on the official KSP forums. In that case, the license reverts
  * back to CC-BY-NC-SA 4.0 INTL.*/
 
-namespace RealChute.Libraries
+namespace RealChute.Libraries.Textures
 {
     public class ModelConfig
     {
