+++ RealChute/RealChute.csproj
@@ -45,47 +45,47 @@
   </ItemGroup>
   <ItemGroup>
     <Compile Include="ChuteTemplate.cs" />
-    <Compile Include="EnumUtils.cs" />
-    <Compile Include="EVAChuteTest.cs" />
+    <Compile Include="Utils\EnumUtils.cs" />
+    <Compile Include="EVA\EVAChuteTest.cs" />
     <Compile Include="Extensions\IEnumerableExtensions.cs" />
     <Compile Include="Extensions\StringBuilderExtensions.cs" />
     <Compile Include="Extensions\VesselExtensions.cs" />
-    <Compile Include="EditorGUI.cs" />
-    <Compile Include="GUIUtils.cs" />
-    <Compile Include="RCToolbarManager.cs" />
-    <Compile Include="IParachute.cs" />
+    <Compile Include="GUI\EditorGUI.cs" />
+    <Compile Include="Utils\GUIUtils.cs" />
+    <Compile Include="Managers\RCToolbarManager.cs" />
+    <Compile Include="Spares\IParachute.cs" />
     <Compile Include="KerbalDumper.cs" />
-    <Compile Include="Libraries\EVAChutes\EVAChute.cs" />
-    <Compile Include="Libraries\EVAChutes\EVAChuteLibrary.cs" />
-    <Compile Include="LinkedToggles.cs" />
-    <Compile Include="ParachuteStorageModule.cs" />
+    <Compile Include="EVA\EVAChute.cs" />
+    <Compile Include="EVA\EVAChuteLibrary.cs" />
+    <Compile Include="GUI\LinkedToggles.cs" />
+    <Compile Include="Spares\ParachuteStorageModule.cs" />
     <Compile Include="PhysicsWatch.cs" />
-    <Compile Include="RealChuteEVA.cs" />
-    <Compile Include="PersistentManager.cs" />
-    <Compile Include="SettingsWindow.cs" />
+    <Compile Include="EVA\RealChuteEVA.cs" />
+    <Compile Include="Managers\PersistentManager.cs" />
+    <Compile Include="GUI\SettingsWindow.cs" />
     <Compile Include="CompatibilityChecker.cs" />
     <Compile Include="Extensions\CelestialBodyExtensions.cs" />
     <Compile Include="Libraries\Presets\Preset.cs" />
     <Compile Include="Libraries\Presets\PresetsLibrary.cs" />
-    <Compile Include="Libraries\TextureLibrary\CanopyConfig.cs" />
-    <Compile Include="Libraries\TextureLibrary\CaseConfig.cs" />
+    <Compile Include="Libraries\Textures\CanopyConfig.cs" />
+    <Compile Include="Libraries\Textures\CaseConfig.cs" />
     <Compile Include="Extensions\ConfigNodeExtensions.cs" />
-    <Compile Include="Libraries\MaterialsLibrary\MaterialDefinition.cs" />
+    <Compile Include="Libraries\Materials\MaterialDefinition.cs" />
     <Compile Include="Extensions\PartExtensions.cs" />
-    <Compile Include="Libraries\MaterialsLibrary\MaterialsLibrary.cs" />
+    <Compile Include="Libraries\Materials\MaterialsLibrary.cs" />
     <Compile Include="Libraries\AtmoPlanets.cs" />
-    <Compile Include="Libraries\TextureLibrary\ModelConfig.cs" />
+    <Compile Include="Libraries\Textures\ModelConfig.cs" />
     <Compile Include="Parachute.cs" />
     <Compile Include="ProceduralChute.cs" />
-    <Compile Include="RCUtils.cs" />
+    <Compile Include="Utils\RCUtils.cs" />
     <Compile Include="RealChuteModule.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="Libraries\TextureLibrary\TextureConfig.cs" />
-    <Compile Include="Libraries\TextureLibrary\TextureLibrary.cs" />
+    <Compile Include="Libraries\Textures\TextureConfig.cs" />
+    <Compile Include="Libraries\Textures\TextureLibrary.cs" />
     <Compile Include="RealChuteSettings.cs" />
     <Compile Include="SizeNode.cs" />
-    <Compile Include="SpareChute.cs" />
-    <Compile Include="TemplateGUI.cs" />
+    <Compile Include="Spares\SpareChute.cs" />
+    <Compile Include="GUI\TemplateGUI.cs" />
   </ItemGroup>
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
   <PropertyGroup>
