+++ RealChute/RealChuteSettings.cs
@@ -1,7 +1,8 @@
 ﻿using System.IO;
 using UnityEngine;
 using RealChute.Extensions;
-using RealChute.Libraries;
+using RealChute.Utils;
+using RealChute.Libraries.Presets;
 
 /* RealChute was made by Christophe Savard (stupid_chris). You are free to copy, fork, and modify RealChute as you see
  * fit. However, redistribution is only permitted for unmodified versions of RealChute, and under attribution clause.
