+++ RealChute/Spares/ParachuteStorageModule.cs
@@ -2,9 +2,11 @@
 using System.Text;
 using System.Collections.Generic;
 using System.Linq;
-using UnityEngine;
-using RealChute.Libraries;
 using RealChute.Extensions;
+using RealChute.EVA;
+using RealChute.GUI;
+using RealChute.Utils;
+using UnityEngine;
 
 /* RealChute was made by Christophe Savard (stupid_chris). You are free to copy, fork, and modify RealChute as you see
  * fit. However, redistribution is only permitted for unmodified versions of RealChute, and under attribution clause.
@@ -15,7 +17,7 @@ using RealChute.Extensions;
  * inactive (no connection) for a period of 90 days on the official KSP forums. In that case, the license reverts
  * back to CC-BY-NC-SA 4.0 INTL.*/
 
-namespace RealChute
+namespace RealChute.Spares
 {
     public enum StorageTab
     {
@@ -232,7 +234,7 @@ namespace RealChute
             EVAChute c = null;
             SpareChute s = null;
             IParachute p = null;
-            GUI.DragWindow(drag);
+            UnityEngine.GUI.DragWindow(drag);
             GUILayout.BeginVertical();
 
             StringBuilder b = new StringBuilder();
